Statistics view button works now
The button "Statistics View" now functions
