+++ Menu_editor/nbproject/private/private.properties
@@ -1,6 +1,10 @@
 compile.on.save=true
 do.depend=false
 do.jar=true
+file.reference.derbyclient-10.9.1.0__1_.jar=C:\\Users\\T D Lo\\Documents\\GitHub\\FatPOS
+file.reference.derbyclient.jar=C:\\Users\\T D Lo\\Documents\\GitHub\\FatPOS
+file.reference.keithburgess-.derby=C:\\Users\\T D Lo\\Documents\\GitHub\\FatPOS
+file.reference.keithburgess-.netbeans-derby=C:\\Users\\T D Lo\\Documents\\GitHub\\FatPOS
 javac.debug=true
 javadoc.preview=true
-user.properties.file=C:\\Users\\Dylan\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
+user.properties.file=C:\\Users\\T D Lo\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
