+++ Menu_editor/nbproject/project.properties
@@ -64,7 +64,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=menu_editor.MenuEditorUI
+main.class=menu_editor.StartMenu_UI
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
